# openapi-complete.yaml
openapi: 3.0.3
info:
  title: AI Guardians API - Unified Guard Processing
  description: |
    **Unified API specification for AI Guardians backend services.**
    
    This API provides comprehensive text analysis through a unified guard processing system.
    All guard services (TokenGuard, TrustGuard, ContextGuard, BiasGuard, HealthGuard) are
    processed through a single unified endpoint with intelligent routing and fallback mechanisms.
    
    ## Unified Architecture
    -  **Unified Processing** - Single endpoint for all guard services
    -  **Intelligent Routing** - Automatic service selection and load balancing
    -  **Fallback Mechanisms** - Automatic failover and service recovery
    -  **Real-time Monitoring** - Health checks and performance metrics
    -  **Enterprise Security** - JWT auth, rate limiting, CORS
    -  **Multi-Client Support** - Web, VS Code, Chrome, API clients
    
    ## Guard Services
    - **TokenGuard** - Token optimization and cost analysis
    - **TrustGuard** - Trust validation and reliability scoring
    - **ContextGuard** - Context analysis and drift detection
    - **BiasGuard** - Bias detection and mitigation
    - **HealthGuard** - System monitoring and security scanning
    
    ## Authentication
    All endpoints (except health) require JWT authentication via Bearer token in Authorization header.
    
    ## Rate Limiting
    - Free Tier: 100 requests/hour
    - Pro Tier: 1000 requests/hour
    - Enterprise: Unlimited
    
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /guards/biasguard/detect:
    post:
      tags: [Guard Services]
      summary: Detect bias
      description: Detect bias in text using BiasGuard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BiasGuardRequest'
      responses:
        '200':
          description: Bias detection completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BiasGuardResponse'
  /guards/trustguard/validate:
    post:
      tags: [Guard Services]
      summary: Validate trust
      description: Validate trust and reliability using TrustGuard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrustGuardRequest'
      responses:
        '200':
          description: Trust validation completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrustGuardResponse'
  /guards/contextguard/analyze:
    post:
      tags: [Guard Services]
      summary: Analyze context
      description: Analyze context and detect drift using ContextGuard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextGuardRequest'
      responses:
        '200':
          description: Context analysis completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextGuardResponse'
  /guards/tokenguard/optimize:
    post:
      tags: [Guard Services]
      summary: Optimize tokens
      description: Optimize token usage and calculate costs using TokenGuard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenGuardRequest'
      responses:
        '200':
          description: Token optimization completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenGuardResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
...
  schemas:
...
    BiasGuardRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          maxLength: 10000
        threshold:
          type: number
          minimum: 0
          maximum: 1
          default: 0.5
        bias_types:
          type: array
          items:
            type: string
            enum: [gender, racial, age, cultural, statistical]
    BiasGuardResponse:
      type: object
      properties:
        score:
          type: number
        detected_biases:
          type: array
          items:
            $ref: '#/components/schemas/DetectedBias'
        suggestions:
          type: array
          items:
            type: string
    TrustGuardRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
        source:
          type: string
        context:
          type: object
    TrustGuardResponse:
      type: object
      properties:
        trust_score:
          type: number
        reliability:
          type: number
        accuracy:
          type: number
        failure_risk:
          type: number
    ContextGuardRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
        conversation_history:
          type: array
          items:
            type: string
    ContextGuardResponse:
      type: object
      properties:
        context_score:
          type: number
        drift_detected:
          type: boolean
        drift_magnitude:
          type: number
        relevance:
          type: number
    TokenGuardRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
        max_tokens:
          type: integer
    TokenGuardResponse:
      type: object
      properties:
        token_count:
          type: integer
        efficiency_score:
          type: number
        cost:
          type: number
        recommendations:
          type: array
          items:
            type: string
...
    DetectedBias:
      type: object
      properties:
        type:
          type: string
          enum: [gender_bias, racial_bias, age_bias, cultural_bias, statistical_bias, semantic_bias]
        confidence:
          type: number
          minimum: 0
          maximum: 1
        location:
          type: object
          properties:
            start:
              type: integer
            end:
              type: integer
        severity:
          type: string
          enum: [low, medium, high, critical]
        suggestion:
          type: string
...